Ajusta tipo da variável de preço

+++ br/com/alura/argentum/modelo/Negociacao.java
@@ -1,20 +1,19 @@
 package br.com.alura.argentum.modelo;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public final class Negociacao {
-    private final BigDecimal preco;
+    private final double preco;
     private final int quantidade;
     private final LocalDateTime data;
 
-    public Negociacao(BigDecimal preco, int quantidade, LocalDateTime data) {
+    public Negociacao(double preco, int quantidade, LocalDateTime data) {
         this.preco = preco;
         this.quantidade = quantidade;
         this.data = data;
     }
 
-    public BigDecimal preco() {
+    public double preco() {
         return this.preco;
     }
